api/pyproject.toml
[project]
name = "supafast"
version = "0.1.0"
description = "supafast - Build APIs instantly"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "email-validator>=2.1.0",
    "fastapi>=0.115.6",
    "phonenumbers>=8.13.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.6",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.36",
    "stripe>=11.4.1",
    "structlog>=24.1.0",
    "supabase>=2.6.0",
    "uvicorn>=0.33.0",
]

[dependency-groups]
dev = [
    "asyncpg-stubs>=0.30.0",
    "black>=24.8.0",
    "flake8>=5.0.4",
    "isort>=5.13.2",
    "mypy>=1.14.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.28.0",
]

[tool.black]
line-length = 79
